+++ BE/chackcheck/src/main/java/sigma/chackcheck/domain/book/BookDetail.java
@@ -1,4 +1,4 @@
-package sigma.chackcheck.domain.bookDetail;
+package sigma.chackcheck.domain.book;
 
 import jakarta.persistence.*;
 import lombok.*;
