+++ BE/chackcheck/src/main/java/sigma/chackcheck/domain/user/User.java
@@ -3,8 +3,7 @@ package sigma.chackcheck.domain.user;
 import jakarta.persistence.*;
 import lombok.*;
 import sigma.chackcheck.domain.bookBorrow.BookBorrow;
-import sigma.chackcheck.domain.bookReserve.BookReserve;
-import sigma.chackcheck.domain.penalty.Penalty;
+import sigma.chackcheck.domain.bookBorrow.BookReserve;
 
 import java.util.List;
